eliminated InserterShortcut legacy class and consolidated to InsertValue
